update 1 file and move 6 files
