+++ frontendMobile/src/screens/student/StudentProfile.jsx
@@ -1,6 +1,6 @@
 import { Image, Text, View, StyleSheet, ScrollView } from "react-native";
-import CircleConquist from '../components/CircleConquist';
-import BottomMenuStudent from '../components/MenuStudent';
+import CircleConquist from '../../components/CircleConquist';
+import BottomMenuStudent from '../../components/MenuStudent';
 
 
 const StudentProfile = () => {
@@ -8,12 +8,12 @@ const StudentProfile = () => {
     return (
         <View style={styles.div_main}>
             <View style={styles.div_perfil}>
-                <Image source={require('../assets/ImgProfile.png')} resizeMode="contain" style={[{ width: 80 }, { height: 80 }]}></Image>
+                <Image source={require('../../assets/ImgProfile.png')} resizeMode="contain" style={[{ width: 80 }, { height: 80 }]}></Image>
                 <View>
                     <Text style={{ fontWeight: 'bold' }}>@Nickname</Text>
                     <Text style={{ fontWeight: 'bold' }}>Nome Sobrenome</Text>
                 </View>
-                <Image source={require('../assets/Icone.png')} style={styles.img_icone} resizeMode="contain"></Image>
+                <Image source={require('../../assets/Icone.png')} style={styles.img_icone} resizeMode="contain"></Image>
             </View>
 
             <View style={styles.div_conquistas}>
@@ -22,9 +22,9 @@ const StudentProfile = () => {
                 </View>
                 <ScrollView style={styles.scrView_carrouselConqs} horizontal={true}>
 
-                    <CircleConquist img={require('../assets/exp_fracao.png')} txt={"Explorador de frações"}></CircleConquist>
-                    <CircleConquist img={require('../assets/camp_deno.png')} txt={"Campeão do Denominador"}></CircleConquist>
-                    <CircleConquist img={require('../assets/estrela_frac.png')} txt={"Estrela fracionária"}></CircleConquist>
+                    <CircleConquist img={require('../../assets/exp_fracao.png')} txt={"Explorador de frações"}></CircleConquist>
+                    <CircleConquist img={require('../../assets/camp_deno.png')} txt={"Campeão do Denominador"}></CircleConquist>
+                    <CircleConquist img={require('../../assets/estrela_frac.png')} txt={"Estrela fracionária"}></CircleConquist>
 
                 </ScrollView>
             </View>
