+++ frontendMobile/src/screens/student/StudentTrails.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { View, Text, ScrollView, StyleSheet, Image, TouchableOpacity } from 'react-native';
-import colors from '../components/styles';
-import BottomMenuStudent from '../components/MenuStudent';
+import colors from '../../components/styles';
+import BottomMenuStudent from '../../components/MenuStudent';
 
 const StudentTrails = ({navigation}) => {
     const handleAccessTrailPress = () => {
@@ -11,31 +11,31 @@ const StudentTrails = ({navigation}) => {
         <View style={styles.container}>
             <View style={styles.divT}>
                 <View style={styles.box}  >
-                    <Image source={require('../assets/trilha.png')} style={styles.iconeBox} />
+                    <Image source={require('../../assets/trilha.png')} style={styles.iconeBox} />
                     <Text style={styles.textBox}>Trilhas</Text>
                 </View>
                 <View style={styles.main}>
                     <ScrollView horizontal={true} style={{ display: 'flex', flexDirection: 'row' }}>
                         <View style={styles.trilha1}>
                             <TouchableOpacity onPress={handleAccessTrailPress}>
-                            <Image source={require('../assets/fracao.png')} style={styles.img} />
+                            <Image source={require('../../assets/fracao.png')} style={styles.img} />
                             <Text style={styles.text}>Trilha de Frações</Text>
                             </TouchableOpacity>
                         </View>
                         <View style={styles.trilha2}>
-                            <Image source={require('../assets/por.png')} style={styles.img} />
+                            <Image source={require('../../assets/por.png')} style={styles.img} />
                             <Text style={styles.text}>Trilha de Porcentagem</Text>
                         </View>
                         <View style={styles.trilha3}>
-                            <Image source={require('../assets/geo.png')} style={styles.img} />
+                            <Image source={require('../../assets/geo.png')} style={styles.img} />
                             <Text style={styles.text}>Trilha de Geometria</Text>
                         </View>
                         <View style={styles.trilha4}>
-                            <Image source={require('../assets/matriz.png')} style={styles.img} />
+                            <Image source={require('../../assets/matriz.png')} style={styles.img} />
                             <Text style={styles.text}>Trilha de matrizes</Text>
                         </View>
                         <View style={styles.trilha5}>
-                            <Image source={require('../assets/expre.png')} style={styles.img} />
+                            <Image source={require('../../assets/expre.png')} style={styles.img} />
                             <Text style={styles.text}>Trilha de Expressão</Text>
                         </View>
                     </ScrollView>
